Create container-with-most-water.java
O(n) TC and O(1) space complexity using two-pointer optimization. The target is to move the pointers with each iteration such that every single potentially larger volume will be examined. This also means that we must move the pointers such that no move will ensure a permanently sub-optimal solution space (one where no bigger solution is possible). With each pointer move inwards, the width gets smaller, so the only way to actually have a larger potential volume is if the height gets bigger. Let's say we have pointer positions [i, j] where h(i) < h(j), and a bigger volume exists in the sub-interval. It could be [i + 1, j] or [i, j + 1]. If we move it to [i, j - 1] then we are absolutely not getting a potentially more optimal solution space since the width of the container is shrinking and the smaller h(i) pole still defines the height of the container that is made. Hence, we must move the left pointer.